frontend/requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.5
aiohttp==3.8.1
httpx==0.23.0
Jinja2==3.0.3
gunicorn==20.1.0